Created function to render text list
Created function to render tex list primitive. Also fix a bug due to scene
pointer not being initialized. To be able to render textlist, the prepare
case was also added. To improve rendering the inline modifier and to let
the compiler handle the optimization. The sharpness of rendering is also
improved by repainting the background every frame.