Implemented symbol table
Implemented symbol table that allow defining and retrieving float variable
or primitive label. Primitive labels are yet to be used and tested. It also
allows pushing and popping table. Popping table is yet to be used and
tested. Also fixed bug with line primitives when direction is left due to
y being decrement wherein it should be x. It doesn't have an obvious effect
but also reassigned primitive on recursive definition of parser for clarity
and to avoid possible bugs.