Implemented keyframes
Implemented keyframes with arbitrary duration. Still have no animations so
easing functions are not implemented yet. Reverted line width to be
constant for now as searching symtable during rendering is incorrect as
it only show that latest val. Also did some profiling and optimization
and the bottleneck seems to be with ffmpeg and did some optimization with
the ffmpeg command.