Implemented last remaining attributes
Implemented last remaining attributes namely from, to, chop, with, and at.
To do this positions are also implemented but rather basic at the moment
as it still doesn't allow for labels. Corners are also implemented but
specifying left and right should come first before top and bot due to
grammar issues that I do not know how to fix.