Implemented '{' and '}' and fix bugs with text
Implemented '{' and '}' which can be used to save states. Also fixed
several bugs with text. First bug is that text list would only show the
last text. And lastly using at attribute with text primitives would not
position properly sometimes because center of the primitive also moves with
text positioning.