Allowed chop attribute to be used on arc primitive
Allowed chop attribute to be used on arc primitive but still some bugs like
when chop value is equal to radius of the arc or if the end of arc and start of arc pass each other. With the first one I do not know the issue, I
think it is due to division by zero or sqrt out of range or domain but I
can't find what part of formula is contributing to it. And with the last
bug if the endpoints pass each other it should invert the counterclockwise
flag.